app/requirements.txt
streamlit
pandas
numpy
requests
psycopg2-binary
pdfplumber
spacy
nltk
bs4
scikit-learn
joblib
wordcloud
streamlit-aggrid
transformers
bertopic
pyarrow
matplotlib
plotly
sacremoses
google-generativeai
